Fix to work with released and updated visual editing package
