+++ simple-cms/nextjs/package.json
@@ -14,8 +14,8 @@
 		"format": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\""
 	},
 	"dependencies": {
-		"@directus/sdk": "^18.0.3",
-		"@directus/visual-editing": "^1.0.0",
+		"@directus/sdk": "^19.1.0",
+		"@directus/visual-editing": "^1.0.1",
 		"@hookform/resolvers": "^3.10.0",
 		"@radix-ui/react-checkbox": "^1.1.3",
 		"@radix-ui/react-collapsible": "^1.1.2",
